+++ engine/sim/profileset_controller.cpp
@@ -1,4 +1,4 @@
-#include "sim_controller.hpp"
+#include "profileset_controller.hpp"
 
 #include "dbc/dbc.hpp"
 #include "dbc/item_set_bonus.hpp"
