+++ source_files/QT_engine.pri
@@ -159,13 +159,13 @@ HEADERS += engine/sim/plot.hpp
 HEADERS += engine/sim/proc.hpp
 HEADERS += engine/sim/proc_rng.hpp
 HEADERS += engine/sim/profileset.hpp
+HEADERS += engine/sim/profileset_controller.hpp
 HEADERS += engine/sim/progress_bar.hpp
 HEADERS += engine/sim/raid_event.hpp
 HEADERS += engine/sim/reforge_plot.hpp
 HEADERS += engine/sim/scale_factor_control.hpp
 HEADERS += engine/sim/sim.hpp
 HEADERS += engine/sim/sim_control.hpp
-HEADERS += engine/sim/sim_controller.hpp
 HEADERS += engine/sim/sim_ostream.hpp
 HEADERS += engine/sim/uptime.hpp
 HEADERS += engine/sim/work_queue.hpp
@@ -351,12 +351,12 @@ SOURCES += engine/sim/plot.cpp
 SOURCES += engine/sim/proc.cpp
 SOURCES += engine/sim/proc_rng.cpp
 SOURCES += engine/sim/profileset.cpp
+SOURCES += engine/sim/profileset_controller.cpp
 SOURCES += engine/sim/progress_bar.cpp
 SOURCES += engine/sim/raid_event.cpp
 SOURCES += engine/sim/reforge_plot.cpp
 SOURCES += engine/sim/scale_factor_control.cpp
 SOURCES += engine/sim/sim.cpp
-SOURCES += engine/sim/sim_controller.cpp
 SOURCES += engine/sim/sim_ostream.cpp
 SOURCES += engine/sim/uptime_benefit.cpp
 SOURCES += engine/util/cache.cpp
