+++ source_files/VS_engine.props
@@ -163,13 +163,13 @@ To change the list of source files run synchronize.py
 		<ClInclude Include="..\engine\sim\proc.hpp" />
 		<ClInclude Include="..\engine\sim\proc_rng.hpp" />
 		<ClInclude Include="..\engine\sim\profileset.hpp" />
+		<ClInclude Include="..\engine\sim\profileset_controller.hpp" />
 		<ClInclude Include="..\engine\sim\progress_bar.hpp" />
 		<ClInclude Include="..\engine\sim\raid_event.hpp" />
 		<ClInclude Include="..\engine\sim\reforge_plot.hpp" />
 		<ClInclude Include="..\engine\sim\scale_factor_control.hpp" />
 		<ClInclude Include="..\engine\sim\sim.hpp" />
 		<ClInclude Include="..\engine\sim\sim_control.hpp" />
-		<ClInclude Include="..\engine\sim\sim_controller.hpp" />
 		<ClInclude Include="..\engine\sim\sim_ostream.hpp" />
 		<ClInclude Include="..\engine\sim\uptime.hpp" />
 		<ClInclude Include="..\engine\sim\work_queue.hpp" />
@@ -354,12 +354,12 @@ To change the list of source files run synchronize.py
 		<ClCompile Include="..\engine\sim\proc.cpp" />
 		<ClCompile Include="..\engine\sim\proc_rng.cpp" />
 		<ClCompile Include="..\engine\sim\profileset.cpp" />
+		<ClCompile Include="..\engine\sim\profileset_controller.cpp" />
 		<ClCompile Include="..\engine\sim\progress_bar.cpp" />
 		<ClCompile Include="..\engine\sim\raid_event.cpp" />
 		<ClCompile Include="..\engine\sim\reforge_plot.cpp" />
 		<ClCompile Include="..\engine\sim\scale_factor_control.cpp" />
 		<ClCompile Include="..\engine\sim\sim.cpp" />
-		<ClCompile Include="..\engine\sim\sim_controller.cpp" />
 		<ClCompile Include="..\engine\sim\sim_ostream.cpp" />
 		<ClCompile Include="..\engine\sim\uptime_benefit.cpp" />
 		<ClCompile Include="..\engine\util\cache.cpp" />
