+++ source_files/cmake_engine.txt
@@ -157,13 +157,13 @@ sim/plot.hpp
 sim/proc.hpp
 sim/proc_rng.hpp
 sim/profileset.hpp
+sim/profileset_controller.hpp
 sim/progress_bar.hpp
 sim/raid_event.hpp
 sim/reforge_plot.hpp
 sim/scale_factor_control.hpp
 sim/sim.hpp
 sim/sim_control.hpp
-sim/sim_controller.hpp
 sim/sim_ostream.hpp
 sim/uptime.hpp
 sim/work_queue.hpp
@@ -348,12 +348,12 @@ sim/plot.cpp
 sim/proc.cpp
 sim/proc_rng.cpp
 sim/profileset.cpp
+sim/profileset_controller.cpp
 sim/progress_bar.cpp
 sim/raid_event.cpp
 sim/reforge_plot.cpp
 sim/scale_factor_control.cpp
 sim/sim.cpp
-sim/sim_controller.cpp
 sim/sim_ostream.cpp
 sim/uptime_benefit.cpp
 util/cache.cpp
